requirements.txt
# Core data and analysis libraries
pandas>=1.5.0
numpy>=1.21.0
yfinance>=0.2.0
scikit-learn>=1.1.0

# Google Sheets integration
gspread>=5.7.0
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0

# Telegram bot
python-telegram-bot>=20.0

# Scheduling
schedule>=1.2.0

# Additional utilities
python-dotenv>=0.19.0
requests>=2.28.0
joblib>=1.2.0

# Optional: For better performance
numba>=0.56.0
